+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Configuration.php
@@ -1107,29 +1107,19 @@ class Configuration implements ConfigurationInterface
                                 ->end()
                             ->end()
                         ->end()
-                        ->arrayNode('serializer')
+                        ->scalarNode('default_serializer')
+                            ->defaultValue('messenger.transport.native_php_serializer')
+                            ->info('Service id to use as the default serializer for the transports.')
+                        ->end()
+                        ->arrayNode('symfony_serializer')
                             ->addDefaultsIfNotSet()
-                            ->beforeNormalization()
-                                ->always()
-                                ->then(function ($config) {
-                                    if (false === $config) {
-                                        return ['id' => null];
-                                    }
-
-                                    if (\is_string($config)) {
-                                        return ['id' => $config];
-                                    }
-
-                                    return $config;
-                                })
-                            ->end()
                             ->children()
-                                ->scalarNode('id')->defaultValue('messenger.transport.native_php_serializer')->end()
-                                ->scalarNode('format')->defaultValue('json')->end()
+                                ->scalarNode('format')->defaultValue('json')->info('Serialization format for the messenger.transport.symfony_serializer service (which is not the serializer used by default).')->end()
                                 ->arrayNode('context')
                                     ->normalizeKeys(false)
                                     ->useAttributeAsKey('name')
                                     ->defaultValue([])
+                                    ->info('Context array for the messenger.transport.symfony_serializer service (which is not the serializer used by default).')
                                     ->prototype('variable')->end()
                                 ->end()
                             ->end()
@@ -1146,6 +1136,7 @@ class Configuration implements ConfigurationInterface
                                 ->fixXmlConfig('option')
                                 ->children()
                                     ->scalarNode('dsn')->end()
+                                    ->scalarNode('serializer')->defaultNull()->info('Service id of a custom serializer to use.')->end()
                                     ->arrayNode('options')
                                         ->normalizeKeys(false)
                                         ->defaultValue([])