+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/ConfigurationTest.php
@@ -325,8 +325,8 @@ class ConfigurationTest extends TestCase
                 'enabled' => !class_exists(FullStack::class) && interface_exists(MessageBusInterface::class),
                 'routing' => [],
                 'transports' => [],
-                'serializer' => [
-                    'id' => 'messenger.transport.native_php_serializer',
+                'default_serializer' => 'messenger.transport.native_php_serializer',
+                'symfony_serializer' => [
                     'format' => 'json',
                     'context' => [],
                 ],