+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/messenger_routing.php
@@ -3,7 +3,7 @@
 $container->loadFromExtension('framework', [
     'serializer' => true,
     'messenger' => [
-        'serializer' => 'messenger.transport.symfony_serializer',
+        'default_serializer' => 'messenger.transport.symfony_serializer',
         'routing' => [
             'Symfony\Component\Messenger\Tests\Fixtures\DummyMessage' => ['amqp', 'audit'],
             'Symfony\Component\Messenger\Tests\Fixtures\SecondMessage' => [