+++ src/Symfony/Component/Messenger/Transport/TransportFactoryInterface.php
@@ -11,6 +11,8 @@
 
 namespace Symfony\Component\Messenger\Transport;
 
+use Symfony\Component\Messenger\Transport\Serialization\SerializerInterface;
+
 /**
  * Creates a Messenger transport.
  *
@@ -20,7 +22,7 @@ namespace Symfony\Component\Messenger\Transport;
  */
 interface TransportFactoryInterface
 {
-    public function createTransport(string $dsn, array $options): TransportInterface;
+    public function createTransport(string $dsn, array $options, SerializerInterface $serializer): TransportInterface;
 
     public function supports(string $dsn, array $options): bool;
 }