+++ src/Symfony/Component/Messenger/CHANGELOG.md
@@ -4,10 +4,12 @@ CHANGELOG
 4.3.0
 -----
 
+ * [BC BREAK] The `TransportFactoryInterface::createTransport()` signature
+   changed: a required 3rd `SerializerInterface` argument was added.
  * Added a new `SyncTransport` along with `ForceCallHandlersStamp` to
    explicitly handle messages synchronously.
  * Added optional parameter `prefetch_count` in connection configuration, 
-   to setup channel prefetch count
+   to setup channel prefetch count.
  * New classes: `RoutableMessageBus`, `AddBusNameStampMiddleware`
    and `BusNameStamp` were added, which allow you to add a bus identifier
    to the `Envelope` then find the correct bus when receiving from